Reorganize gitversion, GitHub Packages, and GitLab registry packages
Move universal helpers into subpackages under kotlin-common-gradle-plugins:
com.huanshankeji.gitversion, com.huanshankeji.github.packages,
com.huanshankeji.gitlab.packageregistry, and com.huanshankeji.maven.

Move Huanshankeji-specific Maven repository content filtering to the team
module. Deduplicate GitHub Packages URL/credential wiring in team helpers.
Register the public-open-source-dependency-repositories settings plugin.

Plugin ids now follow the subpackage layout (breaking within this PR series).
ABI dumps updated; ./gradlew check passes.
