+++ common-gradle-dependencies/build.gradle.kts
@@ -1,6 +1,6 @@
 plugins {
     conventions
-    id("com.huanshankeji.git-version")
+    id("com.huanshankeji.gitversion.git-version")
 }
 
 gitVersion {
