+++ huanshankeji-team-gradle-plugins/src/main/kotlin/com/huanshankeji/team/dokka/github-dokka-convention.gradle.kts
@@ -2,7 +2,7 @@
 
 package com.huanshankeji.team.dokka
 
-import com.huanshankeji.gitCommitHash
+import com.huanshankeji.gitversion.gitCommitHash
 import com.huanshankeji.team.github.defaultRepositoryName
 import com.huanshankeji.team.github.githubRepositoryUrl
 import org.gradle.api.provider.Property
