+++ huanshankeji-team-gradle-plugins/src/main/kotlin/com/huanshankeji/team/maven/HuanshankejiPublicMavenRepositories.kt
@@ -1,18 +1,14 @@
 package com.huanshankeji.team.maven
 
-import com.huanshankeji.DEV_COMMIT_VERSION_REGEX
-import com.huanshankeji.DIRTY_DEV_COMMIT_VERSION_REGEX
-import com.huanshankeji.HUANSHANKEJI_MAVEN_GROUP
-import com.huanshankeji.LEGACY_SNAPSHOT_VERSION_REGEX
-import com.huanshankeji.contentExcludeHuanshankejiNonStableVersions
-import com.huanshankeji.githubPackagesMavenPassword
-import com.huanshankeji.githubPackagesMavenUsername
+import com.huanshankeji.github.packages.githubPackagesMavenRegistrySetUrlAndCredentials
+import com.huanshankeji.gitversion.DEV_COMMIT_VERSION_REGEX
+import com.huanshankeji.gitversion.DIRTY_DEV_COMMIT_VERSION_REGEX
+import com.huanshankeji.gitversion.LEGACY_SNAPSHOT_VERSION_REGEX
 import com.huanshankeji.team.HUANSHANKEJI_IN_LOWERCASE
 import org.gradle.api.Project
 import org.gradle.api.artifacts.dsl.RepositoryHandler
 import org.gradle.api.artifacts.repositories.MavenArtifactRepository
 import org.gradle.kotlin.dsl.maven
-import java.net.URI
 
 /**
  * Configures Maven local (first) and GitHub Packages for [HUANSHANKEJI_MAVEN_GROUP] artifacts.
@@ -35,10 +31,8 @@ fun RepositoryHandler.configurePublicHuanshankejiArtifactRepositories(
     for (repositoryName in githubPackageRepositoryNames) {
         maven {
             name = "GitHubPackages-$repositoryName"
-            url = URI("https://maven.pkg.github.com/$owner/$repositoryName")
-            credentials {
-                username = project.githubPackagesMavenUsername()
-                password = project.githubPackagesMavenPassword()
+            with(project) {
+                githubPackagesMavenRegistrySetUrlAndCredentials(owner, repositoryName)
             }
             content {
                 includeVersionByRegex(HUANSHANKEJI_MAVEN_GROUP, ".*", DEV_COMMIT_VERSION_REGEX)
