+++ huanshankeji-team-gradle-plugins/src/main/kotlin/com/huanshankeji/team/maven/PublicOpenSourceDependencyRepositoriesExtension.kt
@@ -1,15 +1,10 @@
 package com.huanshankeji.team.maven
 
-import com.huanshankeji.contentExcludeHuanshankejiNonStableVersions
-import com.huanshankeji.contentIncludeHuanshankejiDevCommitVersions
-import com.huanshankeji.contentIncludeHuanshankejiDirtyAndLegacySnapshots
-import com.huanshankeji.githubPackagesMavenPassword
-import com.huanshankeji.githubPackagesMavenUsername
+import com.huanshankeji.github.packages.githubPackagesMavenRegistrySetUrlAndCredentials
 import com.huanshankeji.team.HUANSHANKEJI_IN_LOWERCASE
 import org.gradle.api.artifacts.dsl.RepositoryHandler
 import org.gradle.api.artifacts.repositories.MavenArtifactRepository
 import org.gradle.api.initialization.Settings
-import java.net.URI
 
 /**
  * Composable DSL for public OSS repos. No repositories are added unless explicitly configured
@@ -32,10 +27,8 @@ open class PublicOpenSourceDependencyRepositoriesExtension {
         for (repositoryName in repositoryNames) {
             repositories.maven {
                 name = "GitHubPackages-$repositoryName"
-                url = URI("https://maven.pkg.github.com/$owner/$repositoryName")
-                credentials {
-                    username = settings.githubPackagesMavenUsername()
-                    password = settings.githubPackagesMavenPassword()
+                with(settings) {
+                    githubPackagesMavenRegistrySetUrlAndCredentials(owner, repositoryName)
                 }
                 content {
                     contentIncludeHuanshankejiDevCommitVersions()
