+++ huanshankeji-team-gradle-plugins/src/main/kotlin/com/huanshankeji/team/public-open-source-dependency-repositories.settings.gradle.kts
@@ -1,3 +1,5 @@
+package com.huanshankeji.team
+
 import com.huanshankeji.team.maven.PublicOpenSourceDependencyRepositoriesExtension
 
 val extension = extensions.create(
