+++ kotlin-common-gradle-plugins/src/main/kotlin/com/huanshankeji/github/packages/github-packages-maven-publish.gradle.kts
@@ -1,4 +1,4 @@
-package com.huanshankeji
+package com.huanshankeji.github.packages
 
 plugins {
     /*
