+++ kotlin-common-gradle-plugins/src/main/kotlin/com/huanshankeji/gitlab/packageregistry/gitlab-package-registry-project-level-maven-endpoint-publish.gradle.kts
@@ -1,4 +1,6 @@
-package com.huanshankeji
+package com.huanshankeji.gitlab.packageregistry
+
+import com.huanshankeji.maven.publishing
 
 plugins {
     /*
