+++ kotlin-common-gradle-plugins/src/main/kotlin/com/huanshankeji/gitversion/git-version.gradle.kts
@@ -1,6 +1,4 @@
-package com.huanshankeji
-
-import org.gradle.api.provider.Property
+package com.huanshankeji.gitversion
 
 interface GitVersionExtension {
     val baseVersion: Property<String>
