+++ kotlin-common-gradle-plugins/src/main/kotlin/com/huanshankeji/maven-central-publish-conventions.gradle.kts
@@ -1,5 +1,9 @@
 package com.huanshankeji
 
+import com.huanshankeji.gitversion.isDevCommitVersion
+import com.huanshankeji.gitversion.isDirtyDevCommitVersion
+import com.huanshankeji.gitversion.isSnapshotVersion
+
 plugins {
     id("com.vanniktech.maven.publish")
 }
