+++ kotlin-common-gradle-plugins/src/main/kotlin/com/huanshankeji/maven/PublishingExtensions.kt
@@ -1,0 +1,8 @@
+package com.huanshankeji.maven
+
+import org.gradle.api.Action
+import org.gradle.api.Project
+import org.gradle.api.publish.PublishingExtension
+
+fun Project.publishing(configure: Action<PublishingExtension>): Unit =
+    extensions.configure("publishing", configure)
